Refactored joystick utils
